fix floating-point wav reading. patch by Prakash Punnoor <prakash punnoor de>
git-svn-id: https://aften.svn.sourceforge.net/svnroot/aften@15 ef0d8562-5c19-0410-972e-841db63a069c

+++ aften/wav.c
@@ -198,7 +198,7 @@ wavfile_read_samples(WavFile *wf, double *output, int num_samples)
         }
     } else if(bps == 4) {
         if(wf->format == WAVE_FORMAT_IEEEFLOAT) {
-            double *fbuf = (double *)buffer;
+            float *fbuf = (float *)buffer;
 #ifdef WORDS_BIG_ENDIAN
             uint32_t *buf32 = (uint32_t *)buffer;
             for(i=0; i<nr*wf->channels; i++) {
